+++ src/main/java/com/smerp/controller/purchase/GoodsReturnController.java
@@ -179,7 +179,7 @@ public class GoodsReturnController {
 		return plantService.findAll().stream().collect(Collectors.toMap(Plant::getId, Plant::getPlantName));
 	}
 	
-	public Map<Integer, Object> taxCode() {
+	public Map<Double, Object> taxCode() {
 		
 		//return taxCodeRepository.findAllByOrderByTaxCodeAsc().stream().collect(Collectors.toMap(TaxCode::getTaxCode, TaxCode::getTaxCode));
